feat(private-auction): Add private auction endorsement features to UsersService
Introduced methods to check if a user is endorsed in an auction and to update the list of endorsed auctions for a user. These changes enhance the functionality of the UsersService by integrating it with the auction system. Future implementation is required to clean up endorsed auctions when they are deleted or ended.